move log_serv test_serv
